Split files and make yacc more redable
